[Feature Addition] Enabled Player shoot Monster and Bullet be able to disappear.

+++ Source/Game/weapon/weapon.cpp
@@ -23,7 +23,7 @@ void Weapon::DefaultAttack() {
     }
 }
 
-int Weapon::GetCd() {
+double Weapon::GetCd() {
     return cd;
 }
